+++ tests/lib/test_config.py
@@ -23,6 +23,7 @@ class TestFunction(unittest.TestCase):
             'cache_enabled': True,
             'cache_path': None,
             'log_translation': True,
+            'log_content': True,
             'show_notification': True,
             'translation_position': None,
             'column_gap': {